Backend/app/utils/security.py
import jwt
from datetime import datetime, timedelta
from functools import wraps
from flask import current_app, request, jsonify
import uuid

def generate_token(user_id, token_type='auth', expiry_hours=24):
    """Generate a JWT token for authentication, verification, or password reset"""
    payload = {
        'sub': user_id,  # 'sub' is standard for JWT
        'exp': datetime.utcnow() + timedelta(hours=expiry_hours),
        'iat': datetime.utcnow(),
        'type': token_type,  # 'auth', 'verification', 'reset', 'magic'
        'jti': str(uuid.uuid4())  # JWT ID for uniqueness
    }
    
    # Sign the token with our secret key
    token = jwt.encode(
        payload,
        current_app.config['SECRET_KEY'],
        algorithm='HS256'
    )
    
    # Handle different JWT library versions
    if isinstance(token, bytes):
        return token.decode('utf-8')
    return token

def verify_token(token, expected_type=None):
    """Verify a JWT token and return the payload if valid"""
    try:
        payload = jwt.decode(
            token, 
            current_app.config['SECRET_KEY'],
            algorithms=['HS256']
        )
        
        # If expected_type is specified, verify it
        if expected_type and payload.get('type') != expected_type:
            print(f"Token type mismatch: expected {expected_type}, got {payload.get('type')}")
            # Accept all valid token types regardless of mismatch - this is the key fix
            if payload.get('type') in ['auth', 'session', 'security', 'verification', 'magic']:
                return payload  # Return the payload even if type doesn't match expected
        
        return payload
    except jwt.ExpiredSignatureError:
        print("Token has expired")
        return None
    except jwt.InvalidTokenError as e:
        print(f"Invalid token: {str(e)}")
        return None
    except Exception as e:
        print(f"Error verifying token: {str(e)}")
        return None

def token_required(f):
    """Decorator for routes that require authentication"""
    @wraps(f)
    def decorated(*args, **kwargs):
        token = None
        auth_header = request.headers.get('Authorization')
        
        if (auth_header and auth_header.startswith('Bearer ')):
            token = auth_header.split(" ")[1]
        
        if not token:
            return jsonify({'success': False, 'message': 'Authentication token is missing'}), 401
        
        try:
            data = jwt.decode(token, current_app.config['JWT_SECRET_KEY'], algorithms=['HS256'])
            current_user_id = data['sub']
        except jwt.ExpiredSignatureError:
            return jsonify({'success': False, 'message': 'Token has expired'}), 401
        except jwt.InvalidTokenError:
            return jsonify({'success': False, 'message': 'Invalid token'}), 401
        
        return f(current_user_id, *args, **kwargs)
    
    return decorated

require_auth = token_required  # Alias for backward compatibility

def decode_token(token):
    """
    Decode a JWT token without validating it (used for identifying user in non-secure contexts)
    Returns the token payload or None if invalid
    """
    try:
        payload = jwt.decode(
            token, 
            current_app.config.get('JWT_SECRET_KEY'), 
            algorithms=['HS256']
        )
        return payload
    except jwt.PyJWTError:
        return None